Updating on how integrations are counted?
Line 70: **Note:** Integrated API collection with Gitlab is considered as one integration.

Not sure how to put this but we will need to add a point that clarifies that every integration gets added on usage. There is a common confusion like one listed here:

- What counts as '1 integration'? Does that mean per type of system, i.e. Gitlab, Jenkins, Datadog, etc.  Or does it mean per integrated API collection with Gitlab?

This can either be on the main page or based on integrations.